+++ conans/test/functional/editable/layouts_test.py
@@ -5,7 +5,7 @@ import re
 import textwrap
 import unittest
 
-from conans.model.editable_cpp_info import LAYOUTS_FOLDER
+from conans.model.editable_layout import LAYOUTS_FOLDER
 from conans.test.utils.test_files import temp_folder
 from conans.test.utils.tools import TestClient
 from conans.util.files import load, save_files, save
