+++ conans/test/unittests/model/editable_layout/apply_test.py
@@ -7,7 +7,7 @@ import textwrap
 import unittest
 
 from conans.model.build_info import CppInfo
-from conans.model.editable_cpp_info import EditableLayout
+from conans.model.editable_layout import EditableLayout
 from conans.model.ref import ConanFileReference
 from conans.test.utils.test_files import temp_folder
 from conans.util.files import save
