+++ conans/test/unittests/model/editable_layout/parse_test.py
@@ -6,7 +6,7 @@ import textwrap
 import unittest
 
 from conans.client.conf import default_settings_yml
-from conans.model.editable_cpp_info import EditableLayout
+from conans.model.editable_layout import EditableLayout
 from conans.model.options import Options, PackageOptions
 from conans.model.settings import Settings
 from conans.test.utils.test_files import temp_folder
